feat: Case front mod, initial schematics
Front case has been modified to have a cutout for accessing the microSD card from LCD without removing cover. A second cutout was added to give strain relief to the ribbon cable, which is now being used.

KiCad schematics have been added for the indoor section of the  project, though there are several things where the schematics are advanced of the code:
1) The code for the indoor sensor expects a PCF8523 RTC + microSD FeatherWing (Adafruit 2922), not the DS3231 Precision RTC Featherwing (Adafruit 3028) shown.
2) The LCD (Adafruit 5394) is not shown. It connects with a 18 wire ribbon cable to the EYESPI breakout board.
3) The code expects to use the microSD in the PCF8523 Wing, but the change to the DS3231 Wing means there no longer is a microSD reader in the Wing, and the microSD slot in the LCD must be used instead.