feat: Outdoor schematic, documentation
Added schematic for outdoor unit, including solar panel and solar charger that isn't yet part of the code. This is in preparation for the AQI sensor and any other sensors that will require more power than the SHT45 does.

Documentation additions are just to try to get everything documented for this project.