feat: Outdoor json multi-packet Tx
I've modified the Outdoor unit code to send a json message instead of the previous space-separated numbers. While making the overall message very small, it was not condusive to changes in information being sent, or supporting multiple remote units with varying data. Json messages should allow much more flexibility in thus making the code more flexible overall, and easier to add/remove/change information being sent.

The downside of json is the messages are now much bigger, as every number has to have a tag, and extra characters are used. In light of this, I also modified the outdoor code to split the message into multiple sections and send as multiple packets. The indoor code has not yet been modified to accept these packets. The plan is to use a json to hold all of the recieved data instead of the multitude of variables currently used.

TODO:
* Make the message packet splitting more generic, so that it can handle an arbitrary number of packets.
* Perhaps add message x of y to each packet, so if one of the y number of packets doesn't go through, the base station can re-request the one packet that didn't go through? Not sure if that will be worth the effort though.