feat: Outdoor packet X of Y numbering
Outdoor unit code now sends 57 chars per packet. The first char of each packet is the number of the packet. The second char of each packet is the total number of packets being sent. The numbering is the decimal number of the char, NOT the ASCII value of the char!

Typecast the first and second chars of each packet to int to get the packet numbering.