fix: SHT45 read failures after sleep
For a long time the SHT45 has has random failures to get event. Originally I thought this was corrosion on the micro and just added a loop to try again up to 10 times. But a new micro was used and had the same problem. Eventually I figured out I just needed to do a new begin() after it came out of sleep and then it stopped having issues.

To be clear, this wasn't because the micro went to sleep, but because I shut the power to the SHT45 off during sleep to save power. The power back on of the SHT45 is what had random issues with comms afterwards.